requirements.txt
asgiref==3.3.4
certifi==2021.10.8
charset-normalizer==2.0.10
Django==3.2.4
gunicorn==20.1.0
idna==3.3
python-dotenv==0.19.2
pytz==2021.1
requests==2.27.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.8
